pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-agent-spec"
version = "0.1.5"
description = "Open Agent Spec CLI for bootstrapping AI agent projects"
authors = [{ name = "Andrew Whitehouse", email = "andrewswhitehouse@gmail.com" }]
license = { text = "AGPL-3.0-only" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml",
    "rich",
    "behavioral-contracts"
]

[project.scripts]
oas = "oas_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["oas_cli"]

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=oas_cli"
